Its an android application demonstrating the virtual tour of NIT Agartala campus. Different interactive features have been embedded in the application like GeoTagging, shortest possible path between any two chosen locations and so on.
Cyber Tour is an interactive way of visiting a place without being physically present. Cyber tour,
has an increasing amount of usage as the worlds technology getting more advance. With cyber
tour, people can have a view of the actual place like they are actually there and check the path
between any two significant places in campus through devices like computer or phone, without
having to travel to that specific location. Thus it is considered as a convenient alternative to
the actual tour. There are several methods and techniques available to create cyber tours, here
we have used 360 or panoramic tours, android application of this tour which includes many
features such as geotagging, street view, map view etc. The goal of this project is to ease the
hassle of actual campus touring, as the ones who are visiting campus for the first time might
face problem during the process of campus visit. Cyber tour will provide a better platform to
interact with the campus for the students who want to get admission in this college. So for new
students it will be helpful.
Firstly still images and panoromas were taken for each primary location of the campus. Then a
web-portal has been created in which some basic sections like photo gallery, news & events are
displayed. After web portal, android application is created for this cyber tour in which there are
many section such as Geotagging, Street view, shortest path etc. All section of web-portal are
integrated with android app which will direct the user to the cyber tour interface where using
navigation arrows the user can take a tour of the campus.